refactor: changed components' name and location
